chore(@clayui/nav): update Nav component stories
